requirements.txt
pandas
numpy
scikit-learn
streamlit==1.32.0
streamlit-extras
streamlit_js_eval
openai>=1.0.0
azure-cognitiveservices-speech
azure-ai-textanalytics
langchain 
sqlalchemy
import-ipynb
pyodbc
langchain==0.1.11
langchain-experimental
langchain-community
tabulate
langchain_openai
azure-cosmos
docx2txt
pypdf
azure.ai.formrecognizer
requests
python-dotenv
msal
databricks-sql-connector
semantic-kernel==0.9.1b1